requirements.txt
gitignore_parser
jsonargparse[signatures]
matplotlib>=3.4.0
timm
wandb
lightning
transformers
scipy
torch~=2.4.0
torchvision
ipykernel
fvcore
torchmetrics
pycocotools
pandas>=1.3.0
numpy>=1.20.0
pillow>=8.0.0
gradio>=3.50.0
seaborn>=0.11.0